+++ adminer/include/adminer.inc.php
@@ -172,7 +172,7 @@ class Adminer {
 	* @return mixed true for success, string for error message, false for unknown error
 	*/
 	function login(string $login, string $password) {
-		if ($password == "") {
+		if ($password == "" || !password_required()) { // !password_required() - the server would accept any password
 			return lang('Adminer does not support accessing a database without a password, <a href="https://www.adminer.org/en/password/"%s>more information</a>.', target_blank());
 		}
 		return true;